refactor: Add check_fixed_conn_num_shape function for validating sparse operation inputs and enhance type handling
